requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Pillow==10.1.0
python-dotenv==1.0.0
opencv-python==4.9.0.80
numpy==1.26.4
werkzeug==2.3.7
watchdog==2.3
# requests
# lxml
# opencv-python
# numpy
requests
beautifulsoup4
